first draft
First draft which contains two entries of sensor data. The second entry(the one on the bottom) has the "comments" stripped, and is more true to what will be provided. A tool such as (http://allenice.github.io/madoka/editor/) or (https://next.json-generator.com/) may be used to generate and populate the fields.